It's now possible to set certain (system or user-defined) metadata types as public.
Advanced search gained metadata searching abilities.

It's now possible to edit metadata types in user settings, system types can be set public in the admin panel.

The layout of metadata presentation changed slightly - metadata are now displayed directly under a copy row - to be easily connected visually.

The biggest problem in current metadata architecture is that it's impossible to set a system metadata type public on a per-user basis. It's either public for all users or private for all of them.

I'll add a user selection box next to system metadata types in the advanced search UI so that it'd be possible to limit the search to particular user's metadata.

That's about all that can be done in current architecture.
Ticket: #1"